+++ PGS4NET/DisplaySets/DisplayWindow.cs
@@ -21,22 +21,22 @@ public class DisplayWindow
     /// <summary>
     ///     The horizontal offset of the window's top-left corner within the screen.
     /// </summary>
-    public ushort X { get; set; }
+    public int X { get; set; }
 
     /// <summary>
     ///     The vertical offset of the window's top-left corner within the screen.
     /// </summary>
-    public ushort Y { get; set; }
+    public int Y { get; set; }
 
     /// <summary>
     ///     The width of the window.
     /// </summary>
-    public ushort Width { get; set; }
+    public int Width { get; set; }
 
     /// <summary>
     ///     The height of the window.
     /// </summary>
-    public ushort Height { get; set; }
+    public int Height { get; set; }
 
     /// <summary>
     ///     Initializes a new instance with default values.
@@ -60,7 +60,7 @@ public class DisplayWindow
     /// <param name="height">
     ///     The height of the window.
     /// </param>
-    public DisplayWindow(ushort x, ushort y, ushort width, ushort height)
+    public DisplayWindow(int x, int y, int width, int height)
     {
         X = x;
         Y = y;
