+++ PGS4NET/Segments/FinalObjectDefinitionSegment.cs
@@ -43,7 +43,7 @@ public class FinalObjectDefinitionSegment : ObjectDefinitionSegment
     ///     The RLE-compressed data for this portion of the completed object.
     /// </param>
     public FinalObjectDefinitionSegment(PgsTimeStamp pts, PgsTimeStamp dts
-        , VersionedId<ushort> versionedId, byte[] data) : base(pts, dts, versionedId, data)
+        , VersionedId<int> versionedId, byte[] data) : base(pts, dts, versionedId, data)
     {
     }
 }
